[dsch] fix for TypeScript
