+++ TypeScript/ICreateLogger.ts
@@ -1,0 +1,3 @@
+export interface ICreateLogger {
+  (filename: string): Console;
+}
