+++ TypeScript/IFile.ts
@@ -1,0 +1,6 @@
+
+export interface IFile {
+  found: boolean;
+  ext: string;
+  stream: NodeJS.ReadableStream;
+}
