+++ TypeScript/IServeStatic.ts
@@ -1,0 +1,5 @@
+import { IStorage } from "./IStorage";
+
+export interface IServeStatic {
+  (storage: IStorage, logger: Console, port: number): void;
+}
