+++ TypeScript/IStorage.ts
@@ -1,0 +1,5 @@
+import { IFile } from "./IFile";
+
+export interface IStorage {
+  prepare(filename: string): Promise<IFile>;
+}
