+++ TypeScript/logger-alt.d.ts
@@ -1,2 +1,4 @@
+import { ICreateLogger } from "./ICreateLogger";
+
 export declare const createLogger: ICreateLogger;
