+++ TypeScript/logger-alt.js
@@ -4,6 +4,7 @@ const fs = require('node:fs');
 const path = require('node:path');
 const { Console } = require('node:console');
 
+/** @type {(filename: string) => Console} */
 const createLogger = (filename) => {
   const filePath = path.join(process.cwd(), filename);
   const fileStream = fs.createWriteStream(filePath);
