+++ TypeScript/server.js
@@ -2,6 +2,7 @@
 
 const { createServer } = require('node:http');
 
+/** @type {Record<string, string>} */
 const MIME_TYPES = {
   default: 'application/octet-stream',
   html: 'text/html; charset=UTF-8',
@@ -16,15 +17,36 @@ const MIME_TYPES = {
   txt: 'text/plain',
 };
 
+/** @type {(storage: import('./IStorage').IStorage, logger: Console, port: number) => void} */
 const serveStatic = (storage, logger, port) => {
+
+  /** @type {(res: import('node:http').ServerResponse) => void} */
+  const notFound = (res) => {
+    res.writeHead(404, { 'Content-Type': 'text/plain' });
+    res.write('Not Found');
+    res.end();
+  }
+
   const server = createServer(async (req, res) => {
+    if (!req.url) {
+      logger.log(`${req.method} <not-provided> 404`);
+      return notFound(res);
+    }
+
     const file = await storage.prepare(req.url);
-    const statusCode = file.found ? 200 : 404;
+    
+    if (!file.found) {
+      logger.log(`${req.method} ${req.url} 404`);
+      return notFound(res);
+    }
+
     const mimeType = MIME_TYPES[file.ext] || MIME_TYPES.default;
-    res.writeHead(statusCode, { 'Content-Type': mimeType });
+    res.writeHead(200, { 'Content-Type': mimeType });
+
     file.stream.pipe(res);
-    logger.log(`${req.method} ${req.url} ${statusCode}`);
+    logger.log(`${req.method} ${req.url} 200`);
   });
+
   server.listen(port);
 };
