+++ TypeScript/storage.d.ts
@@ -1,15 +1,7 @@
-export interface IFile {
-  found: boolean;
-  ext: string;
-  stream: NodeJS.ReadableStream;
-}
-
-export interface IStorage {
-  constructor(folder: string);
-  prepare(filename: string): Promise<IFile>;
-}
+import { IFile } from "./IFile";
+import { IStorage } from "./IStorage";
 
 export class Storage implements IStorage {
   constructor(folder: string);
-  prepare(filename): Promise<IFile>;
+  prepare(filename: string): Promise<IFile>;
 }
